+++ resources/hypervisor_manager.rb
@@ -1,0 +1,31 @@
+# (c) Copyright 2020 Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+
+OneviewCookbook::ResourceBaseProperties.load(self)
+
+property :new_name, String
+
+default_action :none
+
+action :none do
+end
+
+action :create do
+  OneviewCookbook::Helper.do_resource_action(self, :HypervisorManager, :create_if_missing)
+end
+
+action :update_registration do
+  OneviewCookbook::Helper.do_resource_action(self, :HypervisorManager, :update_registration)
+end
+
+action :delete do
+  OneviewCookbook::Helper.do_resource_action(self, :HypervisorManager, :delete)
+end