+++ spec/unit/resources/matchers_spec.rb
@@ -60,6 +60,9 @@ describe 'oneview_test::default' do
     # oneview_event
     expect(chef_run).to_not create_oneview_event('')
 
+    # oneview_hypervisor_manager
+    expect(chef_run).to_not update_registration_oneview_hypervisor_manager('')
+
     # oneview_fabric
     expect(chef_run).to_not set_oneview_fabric_reserved_vlan_range('')